Refactored so most code is now in modules
