+++ src/config/models.rs
@@ -1,0 +1,7 @@
+
+/// Defines which gpt model to use. Currently set to "gpt-4"
+pub const MODEL: &str = "gpt-4";
+/// Defines default maximum number of tokens available in conversation and response
+pub const MAX_TOKENS: &i32 = &4096;
+/// Defines default temperature of response
+pub const TEMPERATURE: &f32 = &0.6;
